requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
asyncpg==0.31.0
attrs==25.4.0
bandit==1.7.5
bcrypt==4.1.2
bidict==0.23.1
billiard==4.2.4
black==23.12.0
boolean.py==5.0
brotli==1.2.0
CacheControl==0.14.4
celery==5.3.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.13.1
cryptography==46.0.3
cssselect2==0.8.0
cyclonedx-python-lib==4.2.3
defusedxml==0.7.1
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.109.0
filelock==3.20.2
fonttools==4.61.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
ifaddr==0.2.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.3
kombu==5.6.2
license-expression==30.4.4
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
markdown2==2.4.10
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.0
mypy==1.7.1
mypy_extensions==1.1.0
nicegui==1.4.13
numpy==2.4.0
orjson==3.11.5
packageurl-python==0.17.6
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pikepdf==8.10.1
pillow==12.1.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.6.1
platformdirs==4.5.1
pluggy==1.6.0
prometheus-client==0.19.0
prompt_toolkit==3.0.52
propcache==0.4.1
pscript==0.7.7
psycopg2-binary==2.9.9
py-serializable==0.11.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pydyf==0.12.1
Pygments==2.19.2
pyparsing==3.3.1
pyphen==0.17.2
pyproj==3.6.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.0
python-engineio==4.13.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-slugify==8.0.1
python-socketio==5.16.0
pytz==2023.3
PyYAML==6.0.3
redis==5.0.1
requests==2.31.0
rich==14.2.0
rsa==4.9.1
ruff==0.1.8
shapely==2.0.2
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
sqlmodel==0.0.14
starlette==0.35.1
stevedore==5.6.0
text-unidecode==1.3
tinycss2==1.5.1
tinyhtml5==2.0.0
toml==0.10.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.24.0
validate_docbr==1.11.1
vbuild==0.8.2
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.14
weasyprint==67.0
webencodings==0.5.1
websockets==15.0.1
wrapt==2.0.1
wsproto==1.3.2
yarl==1.22.0
zopfli==0.4.0
